backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=1.4.41
psycopg2-binary>=2.9.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0 